refactor: Rename enumSelectModes to enumSelectMode
