+++ user_interface/src/common/enumSelectMode.js
